Edits to ramp math and driver intake controls
Intake controls still not working. Have not gotten a chance to test ramp or debug intake controls. Rick doesn't seem to want driver to be able to control the intakes. Thinking of putting a switch on the intake to automatically deploy the hatch claws when you collide instead. Maybe use 2 switches so you don't accidentally trigger it if you're misaligned?